refactor(JOE-942): extract task-helpers and human-gates domains
Move task/run readiness, create-in-state helpers, and human-gate
list/create/insert into acyclic leaf modules. Ratchet work-store LOC
budget 4500→3950 (~3903 lines).
